python/requirements.txt
# Core dependencies
flask>=3.0.0
flask-cors>=4.0.0

# Speech processing
openai-whisper>=20231117
pyannote.audio>=3.1.1

# Vector database & embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Text-to-speech
TTS>=0.22.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
torch>=2.1.0
torchaudio>=2.1.0
soundfile>=0.12.0

# Audio processing
librosa>=0.10.0
webrtcvad>=2.0.10
